+++ src/aioamazondevices/const/http.py
@@ -31,7 +31,7 @@ CSRF_A2Z = "anti-csrftoken-a2z"
 CSRF_COOKIE = "csrf"
 REQUEST_AGENT = {
     "Amazon": f"AmazonWebView/AmazonAlexa/{AMAZON_APP_VERSION}/iOS/{AMAZON_CLIENT_OS}/iPhone",  # noqa: E501
-    "Browser": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/141.0.0.0 Safari/537.36 Edg/141.0.0.0",  # noqa: E501
+    "Browser": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/152.0.0.0 Safari/537.36 Edg/152.0.0.0",  # noqa: E501
 }
 
 REFRESH_ACCESS_TOKEN = "access_token"  # noqa: S105